closes #299 - splits db seeds into dedicated files
