+++ README.md
@@ -6,7 +6,7 @@ Create shopping lists with items and assemblies (groups of items)
 
 ### Verify Ruby and Node
 
-* `ruby` - 2.6.3
+* `ruby` - 2.6.4
 * `node` - 11.10.1
 
 ### Installation
@@ -16,4 +16,5 @@ $ bundle install
 $ rails db:create
 $ rails db:migrate
 $ rails db:seed
+$ rails start
 ```
