+++ db/seeds/04_departments.rb
@@ -1,0 +1,10 @@
+puts '=== Departments'
+
+department_names = ['Produce', 'Meat', 'Bakery', 'Deli', 'Dairy', 'Fozen', 'Pasta', 'Bread']
+department_names.each do |department_name|
+  Department.create(name: department_name)
+
+  print ','
+end
+
+print "\n\n"
